ms3/requirements.txt
kafka-python
faker
python-dotenv
psycopg2-binary
sqlalchemy
fastapi
uvicorn